+++ internal/services/azapi_resource_action_data_source.go
@@ -1,7 +1,6 @@
 package services
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"log"
@@ -11,30 +10,30 @@ import (
 	"github.com/Azure/terraform-provider-azapi/internal/services/parse"
 	"github.com/Azure/terraform-provider-azapi/internal/services/validate"
 	"github.com/Azure/terraform-provider-azapi/internal/tf"
-	"github.com/Azure/terraform-provider-azapi/utils"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
-func ResourceAzApiOperation() *schema.Resource {
+func ResourceResourceActionDataSource() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceAzApiOperationCreateUpdate,
-		Read:   resourceAzApiOperationRead,
-		Update: resourceAzApiOperationCreateUpdate,
-		Delete: resourceAzApiOperationDelete,
+		Read: resourceResourceActionDataSourceRead,
 
 		Importer: tf.DefaultImporter(func(id string) error {
-			return fmt.Errorf("`azapi_operation` doesn't support import")
+			return fmt.Errorf("`azapi_resource_action` doesn't support import")
 		}),
 
 		Timeouts: &schema.ResourceTimeout{
-			Create: schema.DefaultTimeout(30 * time.Minute),
-			Read:   schema.DefaultTimeout(5 * time.Minute),
-			Update: schema.DefaultTimeout(30 * time.Minute),
-			Delete: schema.DefaultTimeout(30 * time.Minute),
+			Read: schema.DefaultTimeout(30 * time.Minute),
 		},
 
 		Schema: map[string]*schema.Schema{
+			"resource_id": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ForceNew:     true,
+				ValidateFunc: validate.AzureResourceID,
+			},
+
 			"type": {
 				Type:         schema.TypeString,
 				Required:     true,
@@ -42,14 +41,7 @@ func ResourceAzApiOperation() *schema.Resource {
 				ValidateFunc: validate.ResourceType,
 			},
 
-			"resource_id": {
-				Type:         schema.TypeString,
-				Required:     true,
-				ForceNew:     true,
-				ValidateFunc: validate.AzureResourceID,
-			},
-
-			"operation": {
+			"action": {
 				Type:     schema.TypeString,
 				Optional: true,
 				ForceNew: true,
@@ -61,9 +53,7 @@ func ResourceAzApiOperation() *schema.Resource {
 				Default:  "POST",
 				ValidateFunc: validation.StringInSlice([]string{
 					"POST",
-					"PATCH",
-					"PUT",
-					"DELETE",
+					"GET",
 				}, false),
 			},
 
@@ -72,7 +62,6 @@ func ResourceAzApiOperation() *schema.Resource {
 				Optional:         true,
 				ValidateFunc:     validation.StringIsJSON,
 				DiffSuppressFunc: tf.SuppressJsonOrderingDifference,
-				StateFunc:        utils.NormalizeJson,
 			},
 
 			"response_export_values": {
@@ -89,17 +78,10 @@ func ResourceAzApiOperation() *schema.Resource {
 				Computed: true,
 			},
 		},
-
-		CustomizeDiff: func(ctx context.Context, d *schema.ResourceDiff, meta interface{}) error {
-			if d.HasChange("response_export_values") || d.HasChange("operation") {
-				d.SetNewComputed("output")
-			}
-			return nil
-		},
 	}
 }
 
-func resourceAzApiOperationCreateUpdate(d *schema.ResourceData, meta interface{}) error {
+func resourceResourceActionDataSourceRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*clients.Client).ResourceClient
 	ctx, cancel := tf.ForCreateUpdate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
@@ -109,9 +91,10 @@ func resourceAzApiOperationCreateUpdate(d *schema.ResourceData, meta interface{}
 		return err
 	}
 
-	operationName := d.Get("operation").(string)
+	actionName := d.Get("action").(string)
 	method := d.Get("method").(string)
 	body := d.Get("body").(string)
+
 	var requestBody interface{}
 	if len(body) != 0 {
 		err = json.Unmarshal([]byte(body), &requestBody)
@@ -121,21 +104,13 @@ func resourceAzApiOperationCreateUpdate(d *schema.ResourceData, meta interface{}
 	}
 
 	log.Printf("[INFO] request body: %v\n", body)
-	responseBody, err := client.Action(ctx, id.AzureResourceId, operationName, id.ApiVersion, method, requestBody)
+	responseBody, err := client.Action(ctx, id.AzureResourceId, actionName, id.ApiVersion, method, requestBody)
 	if err != nil {
-		return fmt.Errorf("performing action %s of %q: %+v", operationName, id, err)
+		return fmt.Errorf("performing action %s of %q: %+v", actionName, id, err)
 	}
 
 	d.SetId(id.ID())
 	d.Set("output", flattenOutput(responseBody, d.Get("response_export_values").([]interface{})))
 
-	return resourceAzApiOperationRead(d, meta)
-}
-
-func resourceAzApiOperationRead(d *schema.ResourceData, meta interface{}) error {
-	return nil
-}
-
-func resourceAzApiOperationDelete(d *schema.ResourceData, meta interface{}) error {
-	return nil
+	return resourceResourceActionRead(d, meta)
 }