+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.5.0 (Unreleased)
 FEATURES:
-- **New Data Source**: azapi_operation
-- **New Resource**: azapi_operation
+- **New Data Source**: azapi_resource_action
+- **New Resource**: azapi_resource_action
 
 ENHANCEMENTS:
 - Update bicep types to https://github.com/ms-henglu/bicep-types-az/commit/08fe00486aec8ee0e35a7776f81da090c32476de